+++ frontend/src/pages/Profile/Profile.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import {
   User,
   Edit,
@@ -34,6 +34,7 @@ import {
   CreditCard,
   HelpCircle,
 } from "lucide-react";
+import ChatPanel from "./chatPanel";
 
 const ProfilePage = () => {
   // User data state
@@ -174,8 +175,7 @@ const ProfilePage = () => {
   });
   const [showFeedbackForm, setShowFeedbackForm] = useState(null);
   const [activeTab, setActiveTab] = useState("pending");
-  const [activeChat, setActiveChat] = useState(null);
-  const [newMessage, setNewMessage] = useState("");
+  const [activeChatId, setActiveChatId] = useState(null);
   const [showMobileMenu, setShowMobileMenu] = useState(false);
   const [showUserMenu, setShowUserMenu] = useState(false);
 
@@ -293,13 +293,11 @@ const ProfilePage = () => {
     setShowRequestModal(false);
   };
 
-  const sendMessage = () => {
-    if (!newMessage.trim() || !activeChat) return;
-
-    const message = {
+  const handleSendMessage = (messageText) => {
+    const newMessage = {
       id: Date.now(),
       sender: "You",
-      text: newMessage,
+      text: messageText,
       time: new Date().toLocaleTimeString([], {
         hour: "2-digit",
         minute: "2-digit",
@@ -309,29 +307,19 @@ const ProfilePage = () => {
     if (activeTab === "pending") {
       setPendingRequests((prev) =>
         prev.map((req) =>
-          req.id === activeChat
-            ? { ...req, chatMessages: [...req.chatMessages, message] }
+          req.id === activeChatId
+            ? { ...req, chatMessages: [...req.chatMessages, newMessage] }
             : req
         )
       );
     } else {
       setCompletedRequests((prev) =>
         prev.map((req) =>
-          req.id === activeChat
-            ? { ...req, chatMessages: [...req.chatMessages, message] }
+          req.id === activeChatId
+            ? { ...req, chatMessages: [...req.chatMessages, newMessage] }
             : req
         )
       );
-    }
-
-    setNewMessage("");
-  };
-
-  const getActiveRequest = () => {
-    if (activeTab === "pending") {
-      return pendingRequests.find((req) => req.id === activeChat);
-    } else {
-      return completedRequests.find((req) => req.id === activeChat);
     }
   };
 
@@ -339,7 +327,7 @@ const ProfilePage = () => {
   const RequestCard = ({ request, type = "pending" }) => (
     <div
       className={`border rounded-xl p-5 mb-4 transition-all hover:shadow-md cursor-pointer ${
-        activeChat === request.id ? "ring-2 ring-indigo-500" : ""
+        activeChatId === request.id ? "ring-2 ring-indigo-500" : ""
       } ${
         type === "completed"
           ? request.status === "accepted"
@@ -347,7 +335,7 @@ const ProfilePage = () => {
             : "border-red-200 bg-red-50/20"
           : "border-gray-200"
       }`}
-      onClick={() => setActiveChat(request.id)}
+      onClick={() => setActiveChatId(request.id)}
     >
       <div className="flex justify-between">
         <div className="flex items-start gap-3 flex-1">
@@ -549,145 +537,6 @@ const ProfilePage = () => {
       </div>
     </div>
   );
-
-  // Chat Component
-  const ChatPanel = () => {
-    const request = getActiveRequest();
-
-    if (!request) {
-      return (
-        <div className="hidden lg:flex flex-col items-center justify-center h-full bg-gray-50 rounded-lg p-6 border border-gray-200">
-          <MessageSquare size={48} className="text-gray-300 mb-4" />
-          <h3 className="text-lg font-medium text-gray-700 mb-2">
-            Select a conversation
-          </h3>
-          <p className="text-gray-500 text-center">
-            Choose a request from the list to view messages and details
-          </p>
-        </div>
-      );
-    }
-
-    return (
-      <div className="hidden lg:flex flex-col h-full bg-white rounded-lg border border-gray-200">
-        <div className="border-b p-4 flex items-center justify-between">
-          <div className="flex items-center gap-3">
-            <div className="bg-indigo-100 w-10 h-10 rounded-full flex items-center justify-center">
-              <span className="text-indigo-700 font-bold">
-                {request.userPhoto}
-              </span>
-            </div>
-            <div>
-              <h3 className="font-semibold">{request.user}</h3>
-              <div className="flex items-center text-xs text-gray-500">
-                <Star className="text-yellow-500 fill-yellow-500 w-3 h-3 mr-1" />
-                {request.userRating}
-              </div>
-            </div>
-          </div>
-          <button
-            className="p-1 text-gray-500 hover:bg-gray-100 rounded-full"
-            onClick={() => setActiveChat(null)}
-          >
-            <X size={18} />
-          </button>
-        </div>
-
-        <div className="flex-1 overflow-y-auto p-4 space-y-4">
-          <div className="bg-blue-50 rounded-lg p-3 flex items-center">
-            <div className="flex-1">
-              <p className="text-xs text-blue-600">Offering</p>
-              <p className="font-medium">{request.skillOffered}</p>
-            </div>
-            <ArrowRight className="text-blue-400 mx-2" />
-            <div className="flex-1">
-              <p className="text-xs text-blue-600">Requesting</p>
-              <p className="font-medium">{request.skillRequested}</p>
-            </div>
-          </div>
-
-          {request.message && (
-            <div className="bg-gray-50 rounded-lg p-3">
-              <p className="text-sm font-medium text-gray-700">
-                Initial Message
-              </p>
-              <p className="text-gray-600 mt-1">{request.message}</p>
-            </div>
-          )}
-
-          {request.attachment && (
-            <div className="flex items-center text-blue-600 text-sm">
-              <Paperclip size={14} className="mr-1" />
-              <a href="#" className="hover:underline">
-                {request.attachment}
-              </a>
-            </div>
-          )}
-
-          <div className="border-t pt-4 mt-4">
-            <h4 className="text-sm font-medium text-gray-700 mb-3">
-              Conversation
-            </h4>
-
-            {request.chatMessages.length === 0 ? (
-              <p className="text-gray-500 text-center py-8">
-                No messages yet. Start the conversation!
-              </p>
-            ) : (
-              <div className="space-y-3">
-                {request.chatMessages.map((msg) => (
-                  <div
-                    key={msg.id}
-                    className={`flex ${
-                      msg.sender === "You" ? "justify-end" : "justify-start"
-                    }`}
-                  >
-                    <div
-                      className={`max-w-xs md:max-w-md rounded-lg p-3 ${
-                        msg.sender === "You"
-                          ? "bg-indigo-600 text-white"
-                          : "bg-gray-100 text-gray-800"
-                      }`}
-                    >
-                      <p>{msg.text}</p>
-                      <p
-                        className={`text-xs mt-1 ${
-                          msg.sender === "You"
-                            ? "text-indigo-200"
-                            : "text-gray-500"
-                        }`}
-                      >
-                        {msg.time}
-                      </p>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </div>
-
-        <div className="border-t p-3">
-          <div className="flex items-center gap-2">
-            <input
-              type="text"
-              placeholder="Type a message..."
-              className="flex-1 p-2 border rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-transparent"
-              value={newMessage}
-              onChange={(e) => setNewMessage(e.target.value)}
-              onKeyPress={(e) => e.key === "Enter" && sendMessage()}
-            />
-            <button
-              onClick={sendMessage}
-              className="p-2 bg-indigo-600 text-white rounded-lg hover:bg-indigo-700"
-            >
-              <Send size={18} />
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  };
 
   return (
     <div className="min-h-screen bg-gray-50">
@@ -1061,7 +910,7 @@ const ProfilePage = () => {
 
             {/* Requests Section with Tabs */}
             <div className="flex flex-col lg:flex-row gap-6">
-              <div className={`${activeChat ? "lg:w-2/3" : "w-full"}`}>
+              <div className={`${activeChatId ? "lg:w-2/3" : "w-full"}`}>
                 <div className="bg-white rounded-xl shadow-sm p-6 border border-gray-100">
                   <div className="flex justify-between items-center mb-6">
                     <div className="flex space-x-4">
@@ -1177,8 +1026,15 @@ const ProfilePage = () => {
               </div>
 
               {/* Chat Panel */}
-              <div className={`${activeChat ? "lg:w-1/3" : "hidden"}`}>
-                <ChatPanel />
+              <div className={`${activeChatId ? "lg:w-1/3" : "hidden"}`}>
+                <ChatPanel
+                  activeTab={activeTab}
+                  activeChatId={activeChatId}
+                  pendingRequests={pendingRequests}
+                  completedRequests={completedRequests}
+                  onClose={() => setActiveChatId(null)}
+                  onSendMessage={handleSendMessage}
+                />
               </div>
             </div>
           </div>
@@ -1211,46 +1067,44 @@ const ProfilePage = () => {
                   required
                 >
                   <option value="">Select a skill you can offer</option>
-                  {user.skillsOffered.map((skill, index) => (
-                    <option key={index} value={skill}>
+                  {user.skillsOffered.map((skill) => (
+                    <option key={skill} value={skill}>
                       {skill}
                     </option>
                   ))}
                 </select>
               </div>
+
               <div className="mb-4">
                 <label className="block text-sm font-medium mb-1 text-gray-700">
-                  Skill You Need
+                  Skill You're Requesting
                 </label>
-                <select
+                <input
+                  type="text"
                   name="skillRequested"
                   value={newRequest.skillRequested}
                   onChange={handleRequestChange}
                   className="w-full p-3 border rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-transparent"
+                  placeholder="Enter skill you need"
                   required
-                >
-                  <option value="">Select a skill you need</option>
-                  {user.skillsNeeded.map((skill, index) => (
-                    <option key={index} value={skill}>
-                      {skill}
-                    </option>
-                  ))}
-                </select>
+                />
               </div>
+
               <div className="mb-4">
                 <label className="block text-sm font-medium mb-1 text-gray-700">
-                  Your Message
+                  Message
                 </label>
                 <textarea
                   name="message"
                   value={newRequest.message}
                   onChange={handleRequestChange}
-                  rows="4"
                   className="w-full p-3 border rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-transparent"
-                  placeholder="Describe what you're looking for and what you can offer..."
+                  rows="3"
+                  placeholder="Describe what you're looking for..."
                   required
                 />
               </div>
+
               <div className="grid grid-cols-2 gap-4 mb-4">
                 <div>
                   <label className="block text-sm font-medium mb-1 text-gray-700">
@@ -1279,27 +1133,33 @@ const ProfilePage = () => {
                   />
                 </div>
               </div>
+
               <div className="mb-6">
                 <label className="block text-sm font-medium mb-1 text-gray-700">
                   Attachment (Optional)
                 </label>
-                <label className="flex items-center justify-center w-full p-3 border rounded-lg cursor-pointer hover:bg-gray-50">
-                  <Paperclip size={16} className="mr-2 text-gray-500" />
-                  <span className="text-gray-600 truncate max-w-xs">
-                    {newRequest.attachment || "Upload relevant files"}
-                  </span>
+                <label className="flex items-center justify-center w-full p-3 border-2 border-dashed rounded-lg cursor-pointer hover:bg-gray-50">
                   <input
                     type="file"
-                    onChange={handleFileUpload}
                     className="hidden"
+                    onChange={handleFileUpload}
                   />
+                  <div className="text-center">
+                    <Paperclip className="mx-auto text-gray-400" />
+                    <p className="text-sm text-gray-500 mt-1">
+                      {newRequest.attachment
+                        ? newRequest.attachment
+                        : "Click to upload files"}
+                    </p>
+                  </div>
                 </label>
               </div>
+
               <div className="flex justify-end space-x-3">
                 <button
                   type="button"
                   onClick={() => setShowRequestModal(false)}
-                  className="px-4 py-2 border rounded-lg hover:bg-gray-100 text-gray-700"
+                  className="px-4 py-2 border rounded-lg text-gray-700 hover:bg-gray-100"
                 >
                   Cancel
                 </button>
@@ -1307,8 +1167,8 @@ const ProfilePage = () => {
                   type="submit"
                   className="px-4 py-2 bg-indigo-600 text-white rounded-lg hover:bg-indigo-700 flex items-center"
                 >
-                  <Zap size={16} className="mr-1" />
-                  Submit Request
+                  <Zap className="mr-2" size={16} />
+                  Post Request
                 </button>
               </div>
             </form>
